Implement the function plotter

+++ include/diffeq/plotter.h
@@ -3,6 +3,6 @@
 
 #include <stdint.h>
 
-void plot_function(uint16_t width, uint16_t height, uint16_t step, double (*function)(double));
+void plot_function(uint16_t width, uint16_t height, double step, double (*function)(double));
 
 #endif
